cas/cas/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
celery
redis
pandas
openpyxl
django-celery-results